EmbedAppNameInHostTask call EmbedAppNameInHostUtil
